Move getFileCommit to util.go
